add optimization

+++ socialize_main/utils/deleteImage.py
@@ -2,5 +2,6 @@ import os
 
 
 def delete_image(image_url):
-    if os.path.isfile(image_url):
+    if (not image_url is None
+            and os.path.isfile(image_url)):
         os.remove(image_url)
